shorten the get link

+++ actions/admin.ts
@@ -2,12 +2,16 @@
 
 import { currentUser } from "@clerk/nextjs"
 
-import clientPromise, { connectToShortPathCollection } from "@/lib/mongodb"
+import clientPromise from "@/lib/mongodb"
 import { formatBytes, logger } from "@/lib/utils"
 import { DBInfoProps } from "@/types"
 
 const ADMIN_EMAIL = process.env.ADMIN_EMAIL
 
+/**
+ * get database information from both db().admin() and collections
+ * @returns database information
+ */
 export const getDbInfos = async (): Promise<DBInfoProps[] | null> => {
   const user = await currentUser()
   if (!user || user.emailAddresses[0].emailAddress !== ADMIN_EMAIL) return null
@@ -16,7 +20,9 @@ export const getDbInfos = async (): Promise<DBInfoProps[] | null> => {
 
   const dbList = await client.db().admin().listDatabases()
 
-  const filtered = dbList.databases.filter((db) => db.name.includes("@"))
+  const filtered = dbList.databases.filter(
+    (db) => db.name.startsWith("fs_") && db.name.includes("@")
+  )
 
   const info = Promise.all(
     filtered.map(async (db) => ({
@@ -37,14 +43,18 @@ export const getDbInfos = async (): Promise<DBInfoProps[] | null> => {
   return info
 }
 
+/**
+ * remove entire database of select user
+ * @param dbName
+ * @returns
+ */
 export const dropDb = async (dbName: string) => {
   const client = await clientPromise
-  const shortPathCollection = await connectToShortPathCollection()
 
-  await shortPathCollection.deleteMany({ user_id: dbName })
-
-  await client
-    .db(dbName)
-    .dropDatabase()
-    .then(() => logger(`${dbName} [DROPPED!]`))
+  if (dbName.startsWith("fs_") && dbName.includes("@"))
+    return await client
+      .db(dbName)
+      .dropDatabase()
+      .then(() => logger(`${dbName} [DROPPED!]`))
+      .catch((error) => logger(`[REMOVE DATABASE FAILED]${error}`))
 }
